refactor(home): 将home2.vue重命名为home.vue.并移除原有的home.vue
